Rename schema classes to remove 'elyra' from name
